Fix LN by unlink target before relink
